requirements.txt
flask
Pillow
pyttsx3
speechrecognition
transformers
openai
